requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-CORS==5.0.0
mysql-connector-python==9.1.0
gunicorn==22.0.0
eventlet==0.36.1
